Implemented imperial units
Unit system is stored in the user preferences. It can be set when startint the app or in the configuration. The database and the preferences are kept in the metric system to avoid code repetition and abusing of db updates. util.dart contains a class that converts the units from one to another. Consistency is working properly